+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+<a name="1.1.0"></a>
+# 1.1.0 (2016-11-22)
+* Add validator service
+* Add validation for minSessionsInMyAgenda  
+
 <a name="1.0.0"></a>
 # 1.0.0 (2016-11-22)
 * Add config options: allowNextSped, disallowNextSped, minSessionsInMyAgenda